Change the way to use the drawing board. - Since the original way to use the canvas (just hovering to paint) causes the user to start painting as soon as it enters the canvas, it was considered to implement a click-and-drag feature to paint inside the canvas, but it introduced a bug where if the user, not knowing, clicks-and-drags the invisible dot border, it will continue to paint/erase even if the user is no longer holding the mouse click.
Because of both methods having disadvantages, it is that now, the canvas
will be used with a click-to-toggle function, requiring the user to click
inside the canvas to start using it and click again to stop. User can
double click to paint/erase a single dot.

Changes list:
- One click inside the canvas enables/disables the use of the dot (Toggles
useDot true or false).
- Double click paints/erases a single dot in the canvas.
- The mouse-over event, will paint/erase the dot if useDot variable is true.
- If user "leaves" canvas, the useDot variable turns false.
- Edited some comments.